LIO/flask_app/requirements.txt
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.6.2
APScheduler==3.10.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.2
asttokens==2.2.1
attrs==23.2.0
Automat==20.2.0
autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
timedelta
blinker==1.9.0
certifi==2024.2.2
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.4
comm==0.1.2
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==3.4.8
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distro==1.7.0
docstring-to-markdown==0.12
entrypoints==0.4
exceptiongroup==1.2.0
executing==1.2.0
fastjsonschema==2.16.3
flake8==6.0.0
Flask==2.2.5
gunicorn==23.0.0
fonttools==4.39.0
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.31
greenlet==3.1.1
h11==0.14.0
htmlmin==0.1.12
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
ImageHash==4.3.1
importlib-metadata
incremental==21.3.0
ipyaggrid==0.5.2
ipykernel==6.21.3
ipympl==0.9.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.12.0
itables==1.5.4
itsdangerous==2.2.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.2.0
json-lsp==0.2.0
json5==0.9.11
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.17.3
jupyter-server==1.23.6
jupyter-server-mathjax==0.2.6
jupyter-ydoc==0.2.3
jupyter_client==7.4.1
jupyter_core==5.2.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.22.0
keyring==23.5.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
luno-python==0.0.9
MarkupSafe==3.0.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
more-itertools==8.10.0
multimethod==1.9.1
nbclassic==0.5.4
nbclient==0.7.2
nbconvert==7.2.10
nbdime==3.1.1
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.0
outcome==1.3.0.post0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pdfkit==1.0.0
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.4.0
pivottablejs==0.9.0
platformdirs==3.1.0
plotly==5.15.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.10.0
pycparser==2.21
pydantic>=2
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.14.0
PyHamcrest==2.0.2
PyJWT==2.3.0
pylint==2.17.2
pymssql==2.2.7
pyodbc==4.0.39
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.2
python-magic==0.4.24
pytoolconfig==1.2.5
pytz==2022.1
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==25.0.0
reportlab==4.2.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rope==1.7.0
schedule==1.2.1
scikit-learn==1.3.0
scipy==1.9.3
seaborn==0.12.2
SecretStorage==3.3.1
selenium==4.19.0
Send2Trash==1.8.0
service-identity==18.1.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.36
stack-data==0.6.2
statsmodels==0.13.5
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
terminado==0.17.1
textwrap3==0.9.2
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trio==0.25.0
trio-websocket==0.11.1
Twisted==22.1.0
typeguard>=3,<5
typing_extensions==4.11.0
tzlocal==5.2
ujson==5.7.0
uri-template==1.2.0
urllib3==1.26.5
visions==0.7.5
voila==0.4.0
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0
Werkzeug==3.1.3
widgetsnbextension==4.0.5
wrapt==1.15.0
wsproto==1.2.0
ydata-profiling @ https://github.com/pandas-profiling/pandas-profiling/archive/master.zip
ypy-websocket==0.8.2
zipp==1.0.0
zope.interface==5.4.0